Fix for GetDistance so that it correctly calculates all distances and returns the correct value.
git-svn-id: https://subversion.assembla.com/svn/d2bs@452 080cfd25-d5d9-4c59-82ea-70d02a200d8b

+++ trunk/D2Helpers.h
@@ -85,6 +85,6 @@ DWORD __fastcall D2CLIENT_GetMinionCount(UnitAny* pUnit, DWORD dwType);
 VOID __fastcall D2CLIENT_HostilePartyUnit(RosterUnit* pUnit, DWORD dwButton);
 VOID __fastcall D2CLIENT_TakeWaypoint(DWORD dwWaypointId, DWORD dwArea);
 
-double GetDistance(DWORD x1, DWORD y1, DWORD x2, DWORD y2, DistanceType type = Euclidean);
+double GetDistance(long x1, long y1, long x2, long y2, DistanceType type = Euclidean);
 
 #endif
